+++ generate_rdoc.sh
@@ -6,7 +6,7 @@
 # Purpose:      Generates documentation
 #
 # Created:      11th June 2016
-# Updated:      11th June 2016
+# Updated:      13th April 2019
 #
 #############################################################################
 
@@ -15,9 +15,16 @@ rdoc \
 	-x build_gem.sh \
 	-x generate_rdoc.sh \
 	-x run_all_unit_tests.sh \
-	-x test \
-	-x lib/clasp/old_module.rb \
-	-x examples \
 	-x CLASP.gemspec \
+	\
+	-x doc/ \
+	-x gems/ \
+	-x old-gems/ \
+	-x test/scratch/ \
+	\
+	-x ts_all.rb \
+	-x tc_.*\.rb \
+	\
+	$*