+++ test/unit/tc_specifications.rb
@@ -1,4 +1,4 @@
-#! /usr/bin/env ruby
+#!/usr/bin/env ruby
 
 $:.unshift File.join(File.dirname(__FILE__), '../..', 'lib')
 
@@ -8,68 +8,15 @@ require 'test/unit'
 
 class Test_Specifications_1 < Test::Unit::TestCase
 
-	def test_option_with_two_flag_specifications
+	def test_simple_Flag
 
-		specifications	=	[
+		flag_debug	=	CLASP.Flag('--name')
+		flag_logged	=	CLASP.Flag('--logged')
 
-			CLASP.Flag('--action=list', alias: '-l'),
-            CLASP.Flag('--action=change', alias: '-c'),
-			CLASP.Option('--action', alias: '-a'),
-		]
-
-		# With no arguments
-		begin
-
-			argv = []
-			args = CLASP::Arguments.new argv, specifications
-
-			assert_equal 0, args.flags.size
-			assert_equal 0, args.options.size
-			assert_equal 0, args.values.size
-		end
-
-		# With option
-		begin
-
-			argv = %w{ --action=action1 }
-			args = CLASP::Arguments.new argv, specifications
-
-			assert_equal 0, args.flags.size
-			assert_equal 1, args.options.size
-			assert_equal 0, args.values.size
-
-			assert_equal '--action', args.options[0].name
-			assert_equal 'action1', args.options[0].value
-		end
-
-		# With option alias
-		begin
-
-			argv = %w{ -a action2 }
-			args = CLASP::Arguments.new argv, specifications
-
-			assert_equal 0, args.flags.size
-			assert_equal 1, args.options.size
-			assert_equal 0, args.values.size
-
-			assert_equal '--action', args.options[0].name
-			assert_equal 'action2', args.options[0].value
-		end
-
-		# With flag alias
-		begin
-
-			argv = %w{ -c }
-			args = CLASP::Arguments.new argv, specifications
-
-			assert_equal 0, args.flags.size
-			assert_equal 1, args.options.size
-			assert_equal 0, args.values.size
-
-			assert_equal '--action', args.options[0].name
-			assert_equal 'change', args.options[0].value
-		end
-
+		assert_equal flag_debug, flag_debug
+		assert_equal flag_debug, '--name'
+		assert_not_equal flag_logged, flag_debug
 	end
 end
+